ДЗ6.1 WAL-G 1 Резервная копия снимается после развёртывания demo с VM1 и разворачивается на VM2 2 Резервная копия снимается после 30 сек работы приложения с VM1 и разворачивается на VM2. Мелкие улучшения

+++ 06-Углубленное_изучение_бэкапов/hw6.1/upload_data.sh
@@ -5,6 +5,8 @@ export LANGUAGE="C.UTF-8"
 export LC_ALL="C.UTF-8"
 # Загрузка данных в кластер
 
-curl https://edu.postgrespro.ru/demo-20250901-3m.sql.gz -o demo-20250901-3m.sql.gz
+if [ ! -f demo-20250901-3m.sql.gz ]; then
+  curl https://edu.postgrespro.ru/demo-20250901-3m.sql.gz -o demo-20250901-3m.sql.gz
+fi
 
 gunzip -c demo-20250901-3m.sql.gz | sudo -u postgres psql
